webapi/webapi-contactsmanager-sysapps-tests/contactsmanager/ContactsManager_save_categories_multiple.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>ContactsManager Test: ContactsManager_save_photos_categories</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/contacts-manager-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
var t = async_test("Check if save method of ContactsManager can work properly with multiple categories", {timeout: 2000});

t.step(function () {
  assert_true(!!contacts, "The Navigator support contacts");
  contacts.clear().then(function () {/* Do Nothing */}, function () {/* Do Nothing */});

  var mobilePhone = new ContactTelField({
    types: ['home'],
    preferred: true,
    value: '+34698765432'
  });
  var contactName = new ContactName({
    givenNames: ['John'],
    familyNames: ['Doe']
  });
  var contact = new Contact({
    name: contactName,
    phoneNumbers: [mobilePhone],
    categories: ["categories1", "categories2"]
  });
  contacts.save(contact).then(
    function (info) {
      t.step(function () {
        assert_true(!!info, "The Contact initialized");
        assert_equals(info.categories[0], "categories1", "The Contact.categories[0]");
        assert_equals(info.categories[1], "categories2", "The Contact.categories[1]");
      });
      t.done();
    },
    function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  );
});
</script>
